move away from File::MimeInfo and default mime types table
File::MimeInfo requires to use the database on shared-mime-info
which is a problem to install on some systems (MacOS for instance),
because of that, I'm proposing moving over to File::MimeInfo::Simple
and providing a first mime types table for extensions to use in the
same fashion as most of webservers.

Dancer::Util->mime_types provides this table by setting
$SETTINGS->{mime_types} and still allowing the public method
mime_type to be used by the user and overwriting those values at
will. If the mime type is not on the table it'd fall down to
mimetype, from File::MimeInfo::Simple, which tries to work in the
same fashion as mimetype_fu on the Rails world.