Merge route handler registries instead of overwriting them
The previous commit was working when all the routes were defined in
the same file. If routes handlers were defined in more than one file,
when one of them is changed, all the route handlers defined elsewhere
were lost by the refresh.

With this commit, a merge of the route handlers is done instead of
just overwriting the route handler registry.

A test script is added to enlight that behaviour.