Fix the route tree merge engine.
There was a bug in the merge, it's now able to merge correctly two
tress. It uses the old tree as a reference and ovewrites its elements
with the new versions if they exist in the new tree.

Then it adds all new elements present in the new tree and not in the
old one.

That works pretty well but there's still a (minor) bug: we cannot
remove a route handler (the old tree will have it, and this will be
kept).