headers: ARRAYREF instead of HASHREF
The same HTTP header can be sent several time in a response (for instance
'Set-Cookie'). This was not possible when using HASHREF (a key can only
exist once then).

This commit also introduces the `cookies' keyword which gives access
to a singleton representing all cookies to create.

	get '/stuff' => sub {
		cookies->{number} = 42;
	};

This was needed for the session support, as you guess when reading
Dancer::Session::Abstract.