Postpone subsystems' init until all settings are read.
This was a weird and hard to catch bug. By default initializer for
sessions (also templates and logger) are called in the order of keys
in a hash which is completely random. Sometimes secondary settings
like `memcached_servers' were already read and available and sometimes
they were not (for other, less fortunate key names). Besides, these
settings may now be set from $environment.yml config files.