Version 1.121 : Fix POST data parsing
In version 1.120, with the huge refactoring I did
for removing CGI.pm, I broke the POST data parsing.

This fix rely on HTTP::Body to do the hard work.
There's too much complicated and specialized stuff in
there to implement it in Dancer's side.

Moreover, this will make file upload support easier to
implement in the near future.