Refer to cookbook example of checking for logged in user.
In the Dancer::Session docs, there's a simple example of logging in and setting
a session, and checking for the session within a route handler.

Added a note that you probably don't want to have to explicitly check for the
session in every route handler, and to refer to the example in the cookbook
showing the use of a before filter.