Look for plugin settings with lowercased name, too.
Currently, given a plugin named Dancer::Plugin::Foo, we would look for its
settings under the keys 'Dancer::Plugin::Foo' and 'Foo'.

Since most settings in the config file use lowercase names, it makes sense to be
able to put the settings for D::P::Foo under 'foo', so look for that too.