FIX for issue #77
http://github.com/sukria/Dancer/issues#issue/77

This was tricky because if we just fix the issue, we break the route cache
support. Turns out that the cloning that was performed in Dancer::Route was
failing at cloning the $first_match data structure.

After investigating with Clone.pm, it appeared that a real clone did fix the
issue. So I introduced Dancer::Clone to solve the issue without depending on
Clone.pm (which is not in core as far as I can tell).

We should have in mind that this might introduce some performance breakdown
though, and maybe we should consider providing support for a Clone.pm backend
(which is actually written in C).