Add new render_with_layout keyword.
This allows a route handler to generate and return HTML (or other content,
potentially), but still have it rendered within the default layout, or any other
layout.

Basically, template() has changed so that, if you don't supply a view name, it
will look for $options->{content}, and, if found, use that as the content, and
just give that to the layout.  (Or return it directly, if there is no layout
set; wouldn't make much sense to use render_with_layout() then, but, may as well
be thorough :) )

render_with_layout() is syntactical sugar around this new template() behaviour,
so you can say e.g.

  render_with_layout($html);

You can pass tokens to be used within the layout as you would with template(),
and also the options hashref which is passed on to template(), so you could use
a custom layout with e.g.:

  render_with_layout($html, $tokens, { layout => 'layoutname' });

Also includes new tests in t/05_views/03_layout.t to exercise this new
behaviour.