Dancer::Config: add support for "mergeable" settings
Since the configuration comes from two sources (config.yml and the
environment file), there may be conflicts when the same setting
is defined in both files. Currently, the environment takes precedence
and simply overrides the content of any top-level key.

However, in the case of plugins, it's not user-friendly to simply
wipe out all plugin configuration when configuring an extra one
in the environment, nor very scalable to maintain full copies of
the configuration between environments.

For example, having those files:

  config.yml:
    plugins:
      foo:
        bar: baz

  environments/production.yml:
    plugins:
      zlonk:
        bam: kapow

Before this patch, the result would have been (for plugins):

    plugins:
      zlonk:
        bam: kapow

With this patch, the result is:

    plugins:
      foo:
        bar: baz
      zlonk:
        bam: kapow

As expected, if the environment has some configuration for the 'foo'
plugin, the full 'foo' plugin configuration will be overriden.

The implementation allows for other "mergeable" settings to be added
easily.