Dancer::Handler::Wallflower: url to static file handler
The Wallflower handler doesn't start a HTTP server: it simply
reads URL (actually, just the path part) from standard input,
and create local files with the response body that would have
been generated by Dancer (assuming a GET request).

The goal is to be able to write Dancer applications to generate
static web sites that will be served by a vanilla web server.
This is not a generic tool (like HTTrack) to simply offline browsing.