Patch to allow accessor request->route_pattern
This patch allows a Dancer app to know which route pattern won the
request matching.

This can be very useful for introspection in before filters or plugins.

I know we plan to rewrite all the itnernal API and that this patch is not
the best way to do it, but it's very minimalist and mostly harmless.

On the other hand it fills a gap we have regarding basic introspection
and I need this at work. I want to avoid forking Dancer at work, as much as possible.