Dancer::Plugin: improve plugin_setting search
Rationale for the patch:
* $plugin_name was the result of s/// instead of the modified value
* plugin configuration should live in a single, documented, consistent
  place.
* don't return undef if no configuration is available for the plugin

Since for most plugins it is documented to store the config under
plugins / Foo for plugin Dancer::Plugin::Foo, this is the behaviour
we keep.

The previous behaviour was to try in order:
    Foo, Dancer::Plugin::Foo, foo, dancer::plugin::foo

But because of the above bug, the actual value tested were:
    1, Foo, 1, foo

The new behaviour is to check in a single location:
    Dancer::Plugin::Quux -> Quux
    Some::Dancer::Plugin -> Some::Dancer::Plugin

If no config exists, an empty hash is added to the global configuration
hash, which allows to write:

    if( plugin_setting->{foo} ) {
        ...
    }

instead of:

    my $settings = plugin_setting || {};
    if( $settings->{foo} ) {
        ...
    }

The undocumented search for the lowercase version of the name is gone.