added "default" keyword and sanitised auto_page
This change introduces the "default" keyword, that enables defining
a default route action when no defined route matches. This also
leads to sanitising auto_page, which did not play well with routes
like "/whatever" that were matched by the auto_page mechanism.

The Dancer::Route::Registry handling is changed to guarantee that
at most one default is kept in each http method "queue" and also
to ensure that this default appears at the very end of the matching
mechanism.

The change comes along with minimal documentation and test cases.