Fix test suite with new pretty option of the JSON serializer
With previous commits, the JSON serializer uses the pretty flag
in development environement, which happens to be the default one.

So every test get a pretty JSON, with \n and indentations, of course they don't
expect that, and fail.

Now, all these test set the env to production to avoid prettyness of the
serializer.