Added a new constructor "forward".
This constructor does not require an understanding of context to use,
and follows the principles of object orientation and clean code
by using objects to bundle logically connected arguments. Clean code
standards suggest that constructors/functions with more than
3 arguments in a positional list should be avoided. Since there
are only two logical units of information here (the old request and
the new location), the new class method here can reduce the
inputs to two, while preserving module style. Also, use of the
%ENV is eliminated in this section.