Bugfix: check we had a file ext to guess MIME.
When trying to guess a suitable MIME type for a static file being served, make
sure we actually had a file extension.

Otherwise, if requesting a file without an extension, the full path to the
static file is returned as the Content-Type header, as illustrated by squeeks in
https://gist.github.com/860601

If no extension is found, guessing at a content type isn't going to work, so
just return application/data.

Fixes GH-377.