Adding Tie::Hash::NamedCapture to BUILD_REQUIRES.
It's in core since 5.9.5, but if we need it, we should declare it so earlier
Perls can still pass tests, and also to avoid odd fails like:
http://www.cpantesters.org/cpan/report/80e42b9e-7e98-11e0-8484-e88f204d345e

That was a 5.13.11 install, but with no Tie::Hash::NamedCapture, which is odd.

Probably a messed-up install, but we want to avoid spurious fail reports if we
can.