Fix issue #516 - Remove deprecated code in Route.pm
https://github.com/sukria/Dancer/issues/516

When the r() method was provided, a Dancer::Route object could have
two Regexp onject, either one in ->pattern or one in ->regexp.

Now that r() has been completely deprecated, we can clean the code, and only
rely on ->regexp.

that patch removes all the snippets like:

   $self->regexp || $self->pattern

And it alos takes care to do the init stuff accordingly so that we can rely on
->regexp all the time.