Set 'Server' response header as well as 'X-Powered-By'.
Makes sense for Dancer to be shown as the server, if it is.  If the app is
running behind another webserver running as a proxy (Nginx etc), then this won't
be visible unless the server has been configured to pass on that header.

Makes sense for use of Dancer to be visible - who knows, we might even show up
on www.netcraft.com eventually :)