Support the 'PATCH' HTTP verb.
(See RFC5789 for details on PATCH).

This will not work in standalone mode until `HTTP::Server::Simple` has been
updated to support PATCH; I've submittted a pull request for that:

https://github.com/bestpractical/http-server-simple/pull/1

Once the PR has been merged and a new HTTP::Server::Simple release has been
made, we'll need to depend on that version of HTTP::Server::Simple so that the
patch keyword can be used.