Encode log messages properly.
This is a bug discovered by Penfold (thanks!) - if the charset setting was set
to 'UTF-8', as it is in a default scaffolded config.yml, and is_utf8()
returns true, we didn't bother to encode the message.

The problem being, is_utf8() is being used wrongly - in Penfold's words,
"Encode::is_utf8($message) doesn't do what you think it does".  It doesn't mean
it's already encoded to bytes, it means it contains UTF8 chars.

This change should fix the "Wide character in print" errors reported by MiklerGM
when logging data containing UTF8 chars.