Document that the `view` method appends the default template extension.
As this appears to have been a backwards-incompatible change, I think we should
also consider amending the default view() method to check for existence of the
view without appending the default extension, first, then try appending
it if not found.

I'll do that in a PR, as I'd like to run it past the other core devs before it
gets merged.