Add cookies to the response via add_cookie(), not directly.
Don't directly add another Set-Cookie header; just add this cookie to the list
of cookies which will be sent when the response headers are generated.

This means that, if you attempt to set the same cookie multiple times with the
same name, the last value used wins, rather than sending multiple Set-Cookie
headers for the same cookie, which is wrong.

This should fix GH-725.