Add exception type ::Core::Request.
Fixes GH-738.

(Arguably, though, the code that showed up this missing exception type ought to
use Core::Route, I think, as the issue is that `params` was called wrongly from
within a route handler - the problem isn't with the request.  I guess it comes
down to, should the exception type indicate the point the exception was raised,
or the original source of the problem, if known?)