Re-implementing auto-page feature in D::Renderer.
Instead of adding a route to the app to catch requests for /:page and see if we
can handle that, and pass if not, add a new `render_autopage()` to
`Dancer::Renderer` which is called by `Dancer::Handler` if no route / static
file matched.

This immediately makes the auto_page feature more usable, as it can handle
sub-directories properly now (e.g. given a request for `/foo/bar`, a view named
`foo/bar.tt` will be rendered, as you'd expect).

(Heh, look at that - the commit message explaining is almost as long as the code
needed to just fucking do it :) )