Refactor to handle index.$tmpl_ext for "dir index".
If we don't find a direct match for `views/$path`, try appending `/index` to the
path and looking again.

This means that, if the request was for `/foo` and we didn't have
`views/foo`, we'll then look for `views/foo/index.tt`, and use that if found.

This is something I've missed for a long time, and makes the auto_page feature a
lot more useful, I think.