Let the user set the environments to another path
By default, Dancer will load the environment configuration from

    path(setting('appdir'), 'environments')

This change let the user to set the environments in another directory. This is
particulary useful when your configuration is managed outside the application
(like a repository managed by Ops).

A new test file is added to test this feature, and a short documentation about
`DANCER_CONFDIR` and `DANCER_ENVDIR` is added.