this version of wallflower is not maintained
I've started to work again on wallflower, but as a more generic tool
that should work with any PSGI application, not just Dancer. I'll keep
the name, because it's so cool (and a tip of the hat to Dancer).

The goal is still the same: if you manage to generate a list of URLs
for your application, wallflower should turn those into static files,
ready to be pushed to a static website.

I imagine an application where one runs the full framework locally,
and is able to update data (edit blog posts, etc), and then generates
the static web site to be published.

The only thing that the authors of complex applications need to write
is a tool that will list all URLs for the static version of their website.
And that can be refined to output only the "updated" pages, to speed
up generation and upload of the static files.