OpenID snippets for Cookbook.pod
Dancer lacks good examples for how to handle logins. To dive right in, I've added an advanced example "How to offer OpenID logins" to the Cookbook. I hope peer review helps us mature this demo code so we can roll it into an example and hopefully ultimately into a production ready blueprint or a Plugin if enough code can be designed reusable.