"use base" -> "use parent"
'base.pm' is an old module for which usage is deprecated.
'parent.pm' is its recommended replacement and is in the perl core
since 5.10.1. See the perldoc of each module for details.

This patch replace 'use base' with 'use parent' and add 'parent' as
a PREREQ_PM in Makefile.PL for perls older than 5.10.1.