removing cloning of config settings
The concept of fixing the config settings was to fix a problem with
Dancer::Plugin::EscapeHTML (GH #775) which would mess up settings.
This created additional problems (GH #817) and ithas since been
decided to revert the behavior back and let D::P::EscapeHTML try
and sort it down on its own.

Core should help plugins but not at the cost of messing up core
for everyone else.