Make messages from fatal warnings show up in the logs
When running with the `warnings => 1` setting (as is the
default for the development config), these fatal warnings
cause an Error 500 to be returned but do not show up in
the logger at all. This is less than helpful.

This patch changes it so that the warnings show up in the
logger.