tests: Hash randomization can cause .pl MIME to vary
Since Perl 5.18, MIME::Types has started to vary the reported MIME
type for perl scripts. It can be either application/perl or text/perl.
This change will make sure that the test verifies that the
content-type is the same as MIME::Types reports for .pl, without
having to know which one it is.