Add an optional argument to Dancer::Config->load.
Provide configuration explicitly.

This fixes issues where the configuration source is not a YAML file.  (It can be
used in combination with YAML files, but it is not intended to.)  E.g.:

    my $config = from_json($zookeeper->get('/dancer-apps/the-app/config'));
    Dancer::Config->load($config);

This replaces the workaround:

    my $config = from_json($zookeeper->get('/dancer-apps/the-app/config'));
    foreach my $key (keys %{$config}) {
      setting $key => $config->{$key};
    }

which does not work when some settings have hooks.  For instance, you can't put
the logger configuration in there, because you have no guarantee that the "log"
key will be read after the "log_config" key.  When the "log" key is read, the
corresponding logger class is loaded and init'd, but the logger configuration is
sometimes not available yet.

With `load($config)`, we use the normal Dancer config loading mechanisms, which
fire all config hooks *after* the config has been fully loaded.