Test file uploads better, including big files.
Improved the old tests to ensure that the file content received is as supplied
by returning the MD5 hash to compare.

TODO: need to skip the big file upload test if dd isn't available; possibly skip
unless author-testing as creating a *big* file to upload might be rude on
systems with limited space?