Default to text/plain content type...
Previously we defaulted to application/x-www-form-urlencoded if no specific
content type was given.  If we create a HTTP::Body object for that content type,
it'll have no body filehandle to read the body content from.  So, default to
text/plain, but change to application/x-www-form-urlencoded if we had a hashref
of params to turn into a form submission.

I'm not certain this is wise, but it certainly fixes the immediate problem I was
seeing.  If it turns out to be unwise or I find a better way, I'll revert it
happily - but application/x-www-form-urlencoded seems like a fairly arbitrary
guess.