256MB is enough  to demonstrate.
TODO: should this test run only for author testing, seeing as creating large
temp files may  be a bit anti-social and might fail on boxes with very little
space?  Would need to test for existence of dd anyway.  May be better to bin it
now the problem is improved.256MB is enough  to demonstrate.