_read_to_end no longer returns body content
Saving a copy of the body being read and stored into a var that will then go out
of scope for no reason.

The point I've changed here is the only call to _read_to_end in the codebase,
and it doesn't need that raw body copy, just to know that the request has been
read so it can get body params from the HTTP::Body object.