Use different localhost address properly.
Need to tell Test::TCP the address too, and use it eveywhere.

This should, hopefully, stop tests failing on Travis and smokers, when lots of
tests are trying to listen on ports on 127.0.0.1.

It's a half-assed cop-out that doesn't actually fix the possibility of the race
condition, just making it less likely instead, but the alternative involves
rather more work.