Skip session ID validation for D::S::Cookies.
Yes, this makes me feel dirty - a superclass should have no knowledge of its
subclasses, so this is wrong.

On the other hand, the alternative would be breakage for D::S::Cookie users who
update Dancer but not D::S::Cookie, so pragmatically, I think this somewhat
dirty approach is reasonable under the circumstances.