Revert "Merge branch 'bigpresh/bundle_http_body' into bigpresh/1140"
This reverts commit 7a33f70e59eab70c1bd72d75bf91aa0c84ced24e, reversing
changes made to 43bf41bc3084f5cde51ac0a9ef8b76cf969b76b0.

I'd been thinking about bundling a version of HTTP::Body that changes the
handling of request bodies, but it's not a good idea.

Instead, I'll add a config option to allow request->body to keep the raw request
body in RAM as it used to, for anyone who relies on that behaviour, and not
do so for everyone else where memory usage is important.
Conflicts:
	dist.ini
	lib/Dancer/HTTP/Body.pm
	lib/Dancer/HTTP/Body/MultiPart.pm
	lib/Dancer/HTTP/Body/OctetStream.pm
	lib/Dancer/HTTP/Body/UrlEncoded.pm
	lib/Dancer/HTTP/Body/XForms.pm
	lib/Dancer/HTTP/Body/XFormsMultipart.pm

Conflicts:
	dist.ini
	lib/Dancer/HTTP/Body.pm
	lib/Dancer/HTTP/Body/MultiPart.pm
	lib/Dancer/HTTP/Body/OctetStream.pm
	lib/Dancer/HTTP/Body/UrlEncoded.pm
	lib/Dancer/HTTP/Body/XForms.pm
	lib/Dancer/HTTP/Body/XFormsMultipart.pm