Default to the old, store raw body in RAM.
Also, only check the setting when we begin parsing, rather than calling
setting() for every single chunk of data we read from the buffer.

I'm not 100% certain, but I think that the safest bet is to to make
request->body work as expected unless the user sets the config setting to say
"don't store the request body in RAM", rather than change to it not working /
working unpredictably unless you tell it to store it for you.