Tests fail listening on 127.0.0.10.
This had previously been done in PR #1150 to avoid occasional test failures
where the port that Test::TCP picked to listen on ended up in use before it
tried to bind - but that didn't work for *BSD boxes, or, possibly even, any
non-Linux boxes.

It was reverted, but is still here.  Hmm.