Work out which host & port to use ourselves.
In PR #1150, I switched to using 127.0.0.10 rather than 127.0.0.1
for the test server started by Test::TCP to pick a free port on,
because we were getting some spurious test failures due to a race
condition - Test::TCP would pick a free port using Net::EmptyPort, but by the
time the test app had started and tried to bind to it, something else got it
instead.

That doesn't work on FreeBSD and possibly others, thuough - they only have
127.0.0.1/32 on the loopback interface by default, not 127/8.

So, work out for ourselves whether to use an IP unlikely to have any ports in
use if we can, and 127.0.0.1 otherwise, and pick an available port using the
port 0 trick (request port 0, let the OS pick an available port for us).

TODO: this may need to be made more generic and shared with other tests.

TODO: this test is failing - check if it's down to my work here or something
else.

TODO right now: go to bed.