Use more-invalid YAML.
The reason for the failures was that YAML::XS correctly throws an error parsing
'foo: bar: baz', but YAML.pm doesn't.

Now, there's probably a good argument that YAML.pm *should* baulk at it, but it
doesn't, and that's not Dancer's fault.

So, use some more obviously-invalid YAML in our test, and all looks good.