request->address should respect behind_proxy.
If behind_proxy is set, then it makes sense for request->address to return the
address that request->forwarded_for_address() would return.

Otherwise, all code that was written without proxying in mind (and even our
built-in logging, too) will return the IP of the proxy, rather than the client.

TODO: document in the behind_proxy doco that it will change the behaviour of
request->address.

TODO: decide if this is definitely a good idea.  I think it is - I think it's
somewhat surprising that request->address *doesn't* respect behind_proxy.