Test that things fired in the order we expected.
As best I can tell, the original way this test was done, ensuring that some
positive amount of time elapsed between the before_template and after_template
hooks executed was to ensure that both executed, and in the expected order.

However, it seems an odd way to do that, and caused test failures, e.g.:
http://cpantesters.org/cpan/report/4b88f67d-6bfb-1014-be89-d8a126ce55ee
... where the elapsed time was 0, either because of rounding, or low-precision
clocks etc.

Either way - let's just record what happens in what order, then test that
everything we expected to happen happened (and in the order we expected it to).