Test that session var set in hook persisted.
Just for paranoia, make sure that the change to the session made in the after
hook did indeed get persisted into the session storage.

Just in case the session gets flushed by an after hook which fires before the
after hook we used to do the modification - don't want to think that the change
gets silently discarded, which the previous tests I added wouldn't have noticed.