Just use 127.0.0.1 here.
Drop the cleverness of trying to use 127.0.0.11 if we can, to try to avoid race
conditions where the port is in use by the time we go to listen on it.

Lots of our other tests still use 127.0.0.1, so they'd have had the same chances
to fail anyway.

Just let Test::TCP do the work of picking a port for us to listen on, that's
what it's for.

If we do continue to get spurious failures, I'll look in to ways that Test::TCP
could be improved to help avoid them, and failing that, change *all* our tests
to use some cleverer way to get an available port.  It's entirely possible in
fact that it's not really a problem any more, ISTR that Test::TCP used to just
pick a port at random rather than asking for port 0 and letting the OS pick an
available port for you.