Only skip setting cookie if value is as expected.
In some cases, it seems we might change the session ID mid-flight (see e.g.
t/08_session/12_session_name.t for some tests that make sure we can), and that
will fail without this.

On the other hand, in the real world rather than a contrived test case, would
anyone actually do this?