Hold session by session_name and ID.
It's unlikely, but possible, that the session_name could have been changed
during the lifetime of the request - see e.g. t/08_session/12_session_name.t for
tests which test that doing just that works.

Those tests would fail after changing to holding the session in SharedData for
the life of the request because we'd still return the previous session; this
should fix that.