Test to ensure session is actually destroyed.
Had a user report that, since upgrading 1.3500 -> 1.3513,
session->destroy wasn't actually destroying the session, and the user
remained logged in.

Trying to reproduce.

Suspected the caching of the session in Dancer::SharedData to be the
culprit, but that gets reset by Dancer::SharedData->reset_all at the
beginning of each request.