Support SameSite cookie attribute.
Needs careful testing.

Also, we probably don't want to set the attribute if the user hasn't
asked us to; at present the code will set it to 'None' - which, while
that's what a standards-compliant browser will interpret the lack of a
SameSite attribute as, it's probably still better to only send it if we
were told to, and omit it if not?