GH #1239: Fix buggy t/14_serializer/04_request_xml.t
We see a different number of responses in the access file, but there
should always be at least one line. Change the test to account for any
case > 1.

The error report shows that 0 was returned when the test was run
individually. Can't reproduce this locally, so added a TODO item to
potentially investigate later.